+++ Assets/Scripts/Vertex.cs
@@ -6,7 +6,7 @@ namespace LiquidPlanet
     public struct Vertex
     {
         public float3 position, normal;
-        public float4 tangent;
+        public half4 tangent;
         public float2 texCoord0;
     }
